vibrator: do not turn off vibration immediately
If the user, or any app, did setup a 'duration' time for the previous
vibration, we should respect that: read the 'duration' file and, if it
reports a value different than 0, wait that amount of ms before turning
off the vibration.

The 'duration' file does not report the total duration of the vibration,
instead it reports for how many ms the current vibration will continue.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>